% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bandarban_upazilas}
\alias{bandarban_upazilas}
\title{Spatial Data: Bandarban Upazilas}
\format{
An \code{sf} object with multiple polygons and the following variables:
\describe{
\item{name}{Name of the upazila}
\item{geometry}{Spatial polygon data}
\item{...}{Other optional attributes if present}
}
}
\source{
Custom geospatial data curated for the \code{bangladeshPLUS} package.
}
\usage{
data(bandarban_upazilas)
}
\description{
A simple features (\code{sf}) object containing polygon data for the upazilas (sub-districts) of Bandarban district, Bangladesh.
}
\details{
This dataset is used for mapping and analysis of administrative regions in the Bandarban district. It includes the following upazilas:
\itemize{
\item Alikadam
\item Bandarban Sadar
\item Naikkhyongchhari
\item Rowangchhari
\item Lama
}
}
\keyword{datasets}
