% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_plot.R
\name{bd_plot}
\alias{bd_plot}
\title{Plot maps of administrative levels in Bangladesh}
\usage{
bd_plot(level = "country", type = "static")
}
\arguments{
\item{level}{Administrative level: one of "country", "division", "district", "upazila", "union".}

\item{type}{Plot type: "static" or "interactive".}
}
\value{
A tmap object (static or interactive map).
}
\description{
Uses \code{tmap} to visualize different levels of Bangladesh's administrative maps.
}
\examples{
bd_plot("district", "static")

}
