% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bandarban_data.R
\name{get_bandarban_map}
\alias{get_bandarban_map}
\title{Get Filtered Map of Bandarban}
\usage{
get_bandarban_map(upazila = NULL, union = NULL)
}
\arguments{
\item{upazila}{Optional character vector of upazilas to filter. Must be one or more of:
\itemize{
\item{"Alikadam"}
\item{"Bandarban Sadar"}
\item{"Naikkhyongchhari"}
\item{"Rowangchhari"}
\item{"Lama"}
}}

\item{union}{Optional character vector of unions to filter. Must be from the predefined list of unions in Bandarban.}
}
\value{
A filtered \code{sf} object (shapefile) of Bandarban, based on the specified criteria.
}
\description{
Returns a filtered shapefile of Bandarban's administrative regions based on specified upazilas or unions.
}
\examples{
# All upazilas in Bandarban (default)
get_bandarban_map()

# Specific upazilas
get_bandarban_map(upazila = c("Bandarban Sadar", "Lama"))

# Specific unions (across any upazila)
get_bandarban_map(union = c("Bandarban Sadar", "Lama Sadar"))

# Unions within specific upazilas
get_bandarban_map(
  upazila = "Bandarban Sadar",
  union = c("Bandarban Pourasabha", "Suwalok")
)

}
