% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_district}
\alias{map_district}
\title{Bangladesh Administrative Level 2 Shapefile}
\format{
An \code{sf} object with 7 variables:
\describe{
\item{District}{Name of the district (admin level 2)}
\item{ADM2_PCODE}{Administrative level 2 code}
\item{Division}{Name of the division (admin level 1)}
\item{ADM1_PCODE}{Administrative level 1 code}
\item{Country}{Country name (admin level 0)}
\item{ADM0_PCODE}{Administrative level 0 code}
\item{geometry}{MULTIPOLYGON geometry for administrative areas}
}
}
\source{
Bangladesh Bureau of Statistics
}
\usage{
data(map_district)
}
\description{
A spatial dataset containing administrative level 2 boundaries (districts) of Bangladesh.
}
\examples{
data(map_district)
plot(map_district["District"])
}
\keyword{datasets}
